refact : notification test
