+++ tests/GraphQL/Ecosystem/Notifications/NotificationTest.php
@@ -1,0 +1,56 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\GraphQL\Ecosystem\Notifications;
+
+use Tests\TestCase;
+
+class NotificationTest extends TestCase
+{
+    /**
+     * test_notification.
+     *
+     * @return void
+     */
+    public function testNotification()
+    {
+        $response = $this->graphQL(/** @lang GraphQL */ '
+            {
+                notifications(first: 10) {
+                       data {
+                            id,
+                            types {
+                                id
+                            }
+                        },
+                        paginatorInfo {
+                          currentPage
+                          lastPage
+                        }
+                }
+            }
+            ');
+        $this->assertArrayHasKey('data', $response);
+    }
+
+    /**
+     * test_readAll.
+     *
+     * @return void
+     */
+    public function testReadAll()
+    {
+        $response = $this->graphQL(/** @lang GraphQL */ '
+            mutation {
+                readAllNotifications
+            }
+            ');
+        $this->assertArrayHasKey('data', $response);
+        $response->assertJson([
+            'data' => [
+                'readAllNotifications' => true,
+            ],
+        ]);
+    }
+}
